Add files via upload
This is an example of Fibonacci numbers with great numbers of digits, and functions to manipulate them in character format, and plot them in logarithmic graph. The file fibonacci-sorted_v2 contains Fibonacci numbers up to 4000th, and main.py loads data from file, evaluates log10(Fib(n)), as once was made with logarithmic tables, plots the data with matplotlib. It shows exponential increase (linear in log scale)